# api/openapi.yaml
info:
  title: PaySuper payment solution service.
  version: "1.0.0"
  description: Swagger Specification for PaySuper Management API.
servers:
  - url: "https://api.pay.super.com"
    description: Production API
components:
  schemas:
...
    ListProductsResponse:
      type: object
      properties:
        limit:
          description: "@inject_tag: json\" limi\""
          type: integer
        offset:
          description: "@inject_tag: json\" offse\""
          type: integer
        count:
          description: "@inject_tag: json\" coun\""
          type: integer
        items:
          type: array
          items:
            properties:
              id:
                description: "@inject_tag: validate\" omitempty,hexadecimal,len=2\" json\"i\" bson\"_i\""
                type: string
              project_id:
                description: "@inject_tag: validate\" required,hexadecimal,len=2\" json\"project_i\" bson\"project_i\""
                type: string
              object:
                description: "@inject_tag: validate\" require\" json\"objec\""
                type: string
              type:
                description: "@inject_tag: validate\" require\" json\"typ\""
                type: string
              sku:
                description: "@inject_tag: validate\" require\" json\"sk\" bson\"sk\""
                type: string
              default_currency:
                description: "@inject_tag: validate\" required,alpha,len=\" json\"default_currenc\""
                type: string
              enabled:
                description: "@inject_tag: json\" enable\""
                type: boolean
              prices:
                type: array
                items:
                  properties:
                    amount:
                      description: "@inject_tag: json\" amoun\" validate\"required,numeric,gt=\""
                      type: number
                    currency:
                      description: "@inject_tag: json\" currenc\" validate\"omitempty,alpha,len=\""
                      type: string
                    region:
                      description: "@inject_tag: json\" regio\" validate\"omitempty,region_pric\""
                      type: string
                    is_virtual_currency:
                      description: "@inject_tag: json\" is_virtual_currenc\" bson\"is_virtual_currenc\""
                      type: boolean
              created_at:
                type: object
                properties:
                  seconds:
                    description: "9999-12-31T23:59:59Z inclusive."
                    type: integer
                  nanos:
                    description: "inclusive."
                    type: integer
              updated_at:
                type: object
                properties:
                  seconds:
                    description: "9999-12-31T23:59:59Z inclusive."
                    type: integer
                  nanos:
                    description: "inclusive."
                    type: integer
              images:
                description: "@inject_tag: validate\" dive,omitempty,ur\" json\"image\""
                type: array
                items:
                  type: string
              url:
                description: "@inject_tag: validate\" omitempty,ur\" json\"ur\""
                type: string
              pricing:
                description: "@inject_tag: json\" pricin\" validate\"required,oneof=currency manual steam defaul\""
                type: string
              billing_type:
                description: "@inject_tag: json\" billing_typ\" validate\"required,oneof=real virtua\""
                type: string
    ProductPricesResponse:
      type: object
      properties:
        product_price:
          type: array
          items:
            properties:
              amount:
                description: "@inject_tag: json\" amoun\" validate\"required,numeric,gt=\""
                type: number
              currency:
                description: "@inject_tag: json\" currenc\" validate\"omitempty,alpha,len=\""
                type: string
              region:
                description: "@inject_tag: json\" regio\" validate\"omitempty,region_pric\""
                type: string
              is_virtual_currency:
                description: "@inject_tag: json\" is_virtual_currenc\" bson\"is_virtual_currenc\""
                type: boolean
    Product:
      type: object
      properties:
        id:
          description: "@inject_tag: validate\" omitempty,hexadecimal,len=2\" json\"i\" bson\"_i\""
          type: string
        project_id:
          description: "@inject_tag: validate\" required,hexadecimal,len=2\" json\"project_i\" bson\"project_i\""
          type: string
        object:
          description: "@inject_tag: validate\" require\" json\"objec\""
          type: string
        type:
          description: "@inject_tag: validate\" require\" json\"typ\""
          type: string
        sku:
          description: "@inject_tag: validate\" require\" json\"sk\" bson\"sk\""
          type: string
        default_currency:
          description: "@inject_tag: validate\" required,alpha,len=\" json\"default_currenc\""
          type: string
        enabled:
          description: "@inject_tag: json\" enable\""
          type: boolean
        prices:
          type: array
          items:
            properties:
              amount:
                description: "@inject_tag: json\" amoun\" validate\"required,numeric,gt=\""
                type: number
              currency:
                description: "@inject_tag: json\" currenc\" validate\"omitempty,alpha,len=\""
                type: string
              region:
                description: "@inject_tag: json\" regio\" validate\"omitempty,region_pric\""
                type: string
              is_virtual_currency:
                description: "@inject_tag: json\" is_virtual_currenc\" bson\"is_virtual_currenc\""
                type: boolean
        created_at:
          type: object
          properties:
            seconds:
              description: "9999-12-31T23:59:59Z inclusive."
              type: integer
            nanos:
              description: "inclusive."
              type: integer
        updated_at:
          type: object
          properties:
            seconds:
              description: "9999-12-31T23:59:59Z inclusive."
              type: integer
            nanos:
              description: "inclusive."
              type: integer
        images:
          description: "@inject_tag: validate\" dive,omitempty,ur\" json\"image\""
          type: array
          items:
            type: string
        url:
          description: "@inject_tag: validate\" omitempty,ur\" json\"ur\""
          type: string
        pricing:
          description: "@inject_tag: json\" pricin\" validate\"required,oneof=currency manual steam defaul\""
          type: string
        billing_type:
          description: "@inject_tag: json\" billing_typ\" validate\"required,oneof=real virtua\""
          type: string
    ProductPrice:
      type: object
      properties:
        amount:
          description: "@inject_tag: json\" amoun\" validate\"required,numeric,gt=\""
          type: number
        currency:
          description: "@inject_tag: json\" currenc\" validate\"omitempty,alpha,len=\""
          type: string
        region:
          description: "@inject_tag: json\" regio\" validate\"omitempty,region_pric\""
          type: string
        is_virtual_currency:
          description: "@inject_tag: json\" is_virtual_currenc\" bson\"is_virtual_currenc\""
          type: boolean
...
    GetProductResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: object
          properties:
            code:
              description: "The error code."
              type: string
            message:
              description: "The error short description."
              type: string
            details:
              description: "The error details."
              type: string
        item:
          type: object
          properties:
            id:
              description: "@inject_tag: validate\" omitempty,hexadecimal,len=2\" json\"i\" bson\"_i\""
              type: string
            project_id:
              description: "@inject_tag: validate\" required,hexadecimal,len=2\" json\"project_i\" bson\"project_i\""
              type: string
            object:
              description: "@inject_tag: validate\" require\" json\"objec\""
              type: string
            type:
              description: "@inject_tag: validate\" require\" json\"typ\""
              type: string
            sku:
              description: "@inject_tag: validate\" require\" json\"sk\" bson\"sk\""
              type: string
            default_currency:
              description: "@inject_tag: validate\" required,alpha,len=\" json\"default_currenc\""
              type: string
            enabled:
              description: "@inject_tag: json\" enable\""
              type: boolean
            prices:
              type: array
              items:
                properties:
                  amount:
                    description: "@inject_tag: json\" amoun\" validate\"required,numeric,gt=\""
                    type: number
                  currency:
                    description: "@inject_tag: json\" currenc\" validate\"omitempty,alpha,len=\""
                    type: string
                  region:
                    description: "@inject_tag: json\" regio\" validate\"omitempty,region_pric\""
                    type: string
                  is_virtual_currency:
                    description: "@inject_tag: json\" is_virtual_currenc\" bson\"is_virtual_currenc\""
                    type: boolean
            created_at:
              type: object
              properties:
                seconds:
                  description: "9999-12-31T23:59:59Z inclusive."
                  type: integer
                nanos:
                  description: "inclusive."
                  type: integer
            updated_at:
              type: object
              properties:
                seconds:
                  description: "9999-12-31T23:59:59Z inclusive."
                  type: integer
                nanos:
                  description: "inclusive."
                  type: integer
            images:
              description: "@inject_tag: validate\" dive,omitempty,ur\" json\"image\""
              type: array
              items:
                type: string
            url:
              description: "@inject_tag: validate\" omitempty,ur\" json\"ur\""
              type: string
            pricing:
              description: "@inject_tag: json\" pricin\" validate\"required,oneof=currency manual steam defaul\""
              type: string
            billing_type:
              description: "@inject_tag: json\" billing_typ\" validate\"required,oneof=real virtua\""
              type: string
...
    ResponseErrorMessage:
      type: object
      properties:
        code:
          description: "The error code."
          type: string
        message:
          description: "The error short description."
          type: string
        details:
          description: "The error details."
          type: string
...
paths:
...
  /admin/api/v1/products:
    get:
      summary: Get the list of products
      operationId: productsPathGetProductsList
      description: Get the list of products for the authorised user
      tags:
      - Product
      parameters:
      - name: name
        description: "The product's name."
        in: query
        required: false
        schema:
          type: string
      - name: sku
        description: "The product's SKU."
        in: query
        required: false
        schema:
          type: string
      - name: project_id
        description: The unique identifier for the project.
        in: query
        required: false
        schema:
          type: string
      - name: enabled
        description: "The status of whether the product is enabled. Available values: all, true, false."
        in: query
        required: false
        schema:
          type: string
      - name: limit
        description: The number of products returned in one page. Default value is 100.
        in: query
        required: true
        schema:
          type: string
      - name: offset
        description: The ranking number of the first item on the page.
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "Returns the list of products for the authorised user. This list can be filtered by the product's name, SKU, status and the project ID."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProductsResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
    post:
      summary: Create a product
      operationId: productsPathCreateProduct
      description: Create a new product in the project
      tags:
      - Product
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Returns the created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /system/api/v1/products/merchant/{merchant_id}:
    get:
      summary: Get the list of products using the merchant ID
      operationId: productsMerchantPathGetProductsList
      description: Get the list of products using the merchant ID
      tags:
      - Product
      parameters:
      - name: merchant_id
        description: The unique identifier for the merchant.
        in: path
        required: true
        schema:
          type: string
      - name: name
        description: "The product's name."
        in: query
        required: false
        schema:
          type: string
      - name: sku
        description: "The product's SKU."
        in: query
        required: false
        schema:
          type: string
      - name: project_id
        description: The unique identifier for the project.
        in: query
        required: false
        schema:
          type: string
      - name: enabled
        description: "The status of whether the product is enabled. Available values: all, true, false."
        in: query
        required: false
        schema:
          type: string
      - name: limit
        description: The number of products returned in one page. Default value is 100.
        in: query
        required: true
        schema:
          type: string
      - name: offset
        description: The ranking number of the first item on the page.
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "Returns the list of merchant's products. This list can be filtered by the product's name, SKU, status and the project ID."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProductsResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /admin/api/v1/products/{product_id}:
    get:
      summary: Get the product
      operationId: productsIdPathGetProduct
      description: Get the product using the product ID
      tags:
      - Product
      parameters:
      - name: product_id
        description: The unique identifier for the product.
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Returns the product data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProductResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
    delete:
      summary: Delete the product
      operationId: productsIdPathDeleteProduct
      description: Delete the product using the product ID
      tags:
      - Product
      parameters:
      - name: product_id
        description: The unique identifier for the product.
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Returns an empty response body if the product was successfully removed
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
    put:
      summary: Update the product
      operationId: productsIdPathUpdateProduct
      description: Update the product using the product ID
      tags:
      - Product
      parameters:
      - name: product_id
        description: The unique identifier for the product.
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Returns the updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /admin/api/v1/products/{product_id}/prices:
    get:
      summary: "Get the product's prices"
      operationId: productsPricesPathGetProductPrices
      description: "Get the product's prices using the product ID"
      tags:
      - Product
      parameters:
      - name: product_id
        description: The unique identifier for the product.
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Returns the list of the product's prices"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPricesResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
    put:
      summary: "Set the product's price"
      operationId: productsPricesPathUpdateProductPrices
      description: "Set the product's price using the product ID"
      tags:
      - Product
      parameters:
      - name: product_id
        description: The unique identifier for the product.
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ProductPrice"
      responses:
        "200":
          description: "Returns an empty response body if the product's price was successfully set"
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
...
openapi: "3.0.2"